#5-refactor: messenger 관련 entity, service, controller, repository 파일을 각각 messenger 디렉토리로 이동하여 파일 구조 변경
