+++ src/main/java/com/example/demo/entity/messenger/ChatRoom.java
@@ -1,4 +1,4 @@
-package com.example.demo.entity;
+package com.example.demo.entity.messenger;
 
 import jakarta.persistence.*;
 import lombok.*;
