+++ src/main/java/com/example/demo/entity/messenger/Message.java
@@ -1,4 +1,4 @@
-package com.example.demo.entity;
+package com.example.demo.entity.messenger;
 
 import com.example.demo.entity.users.user.User;
 import jakarta.persistence.*;
