+++ src/main/java/com/example/demo/entity/messenger/UnreadMemberId.java
@@ -1,6 +1,5 @@
-package com.example.demo.entity;
+package com.example.demo.entity.messenger;
 
-import jakarta.persistence.*;
 import lombok.*;
 
 import java.io.Serializable;
