+++ src/main/java/com/example/demo/repository/messenger/MessageRepository.java
@@ -1,7 +1,7 @@
-package com.example.demo.repository;
+package com.example.demo.repository.messenger;
 
 
-import com.example.demo.entity.Message;
+import com.example.demo.entity.messenger.Message;
 import com.example.demo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
