+++ src/main/java/com/example/demo/repository/messenger/UnreadMemberRepository.java
@@ -1,8 +1,7 @@
-package com.example.demo.repository;
+package com.example.demo.repository.messenger;
 
 
-import com.example.demo.entity.Message;
-import com.example.demo.entity.UnreadMember;
+import com.example.demo.entity.messenger.UnreadMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UnreadMemberRepository extends JpaRepository<UnreadMember,Long> {
